fix deployment

+++ travis-deployment/deploy-to-gh-pages.sh
@@ -1,14 +1,25 @@
-#!/bin/bash
+#!/usr/bin/env bash
+set -euo pipefail
+
+echo decrypting the ssh key ...
+
+openssl aes-256-cbc \
+  -K "$encrypted_51a4a26b95d4_key" -iv "$encrypted_51a4a26b95d4_iv" \
+  -in deploy-key.enc -out ~/.ssh/id_rsa -d
+chmod 0700 ~/.ssh
+chmod 0600 ~/.ssh/id_rsa
+
 echo "start deployment ..."
-set -e
+
+GH_REF=git@github.com:nixcon/2019.nixcon.org.git
 
 TMP_CHECKOUT_DIR=tmp/origin-gh-pages
 (
   set -e
-  git clone --branch=gh-pages "https://${GH_TOKEN}@${GH_REF}" $TMP_CHECKOUT_DIR
+  git clone --branch=gh-pages "$GH_REF" $TMP_CHECKOUT_DIR
 
   echo "deploy: checked out 'gh-pages' branch"
-  rm -rf $TMP_CHECKOUT_DIR/*
+  rm -rf ${TMP_CHECKOUT_DIR:?}/*
   cp -r result/* $TMP_CHECKOUT_DIR
   cp CNAME $TMP_CHECKOUT_DIR
   cd $TMP_CHECKOUT_DIR
@@ -20,8 +31,8 @@ TMP_CHECKOUT_DIR=tmp/origin-gh-pages
   git add --all .
 
   echo "deploy: committing"
-  git commit -m "Travis deployed 'master' - `date`" || true
+  git commit -m "Travis deployed 'master' - $(date)" || true
 
   echo "deploy: push to gh-pages"
-  git push "https://${GH_TOKEN}@${GH_REF}" gh-pages:gh-pages
+  git push origin gh-pages:gh-pages
 )
